wakanda-protocol/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
tenacity==8.2.3
pyyaml==6.0.1
cachetools==5.3.2
scikit-learn==1.3.2
pandas==2.1.4
joblib==1.3.2
hvac==2.0.0
boto3==1.34.0
azure-identity==1.15.0
gunicorn==21.2.0
prometheus-client==0.19.0